fix(openai): preserve streaming audio deltas
