+++ src/core/providers/base/sse/cohere.rs
@@ -84,6 +84,7 @@ impl SSETransformer for CohereTransformer {
                                 thinking: None,
                                 tool_calls: None,
                                 function_call: None,
+                                audio: None,
                             },
                             finish_reason: None,
                             logprobs: None,
@@ -136,6 +137,7 @@ impl SSETransformer for CohereTransformer {
                                 thinking: None,
                                 tool_calls: None,
                                 function_call: None,
+                                audio: None,
                             },
                             finish_reason: Some(Self::parse_cohere_finish_reason(finish_reason)),
                             logprobs: None,
@@ -165,6 +167,7 @@ impl SSETransformer for CohereTransformer {
                                 thinking: None,
                                 tool_calls: None,
                                 function_call: None,
+                                audio: None,
                             },
                             finish_reason: None,
                             logprobs: None,
@@ -191,6 +194,7 @@ impl SSETransformer for CohereTransformer {
                                 thinking: None,
                                 tool_calls: None,
                                 function_call: None,
+                                audio: None,
                             },
                             finish_reason: Some(Self::parse_cohere_finish_reason(finish_reason)),
                             logprobs: None,
