+++ src/core/providers/langgraph/provider.rs
@@ -670,6 +670,7 @@ fn parse_langgraph_sse_event(event: &str, model: &str) -> Option<ChatChunk> {
                                 thinking: None,
                                 tool_calls: None,
                                 function_call: None,
+                                audio: None,
                             },
                             finish_reason: None,
                             logprobs: None,
@@ -695,6 +696,7 @@ fn parse_langgraph_sse_event(event: &str, model: &str) -> Option<ChatChunk> {
                         thinking: None,
                         tool_calls: None,
                         function_call: None,
+                        audio: None,
                     },
                     finish_reason: Some(crate::core::types::responses::FinishReason::Stop),
                     logprobs: None,
