+++ src/core/types/responses/chat.rs
@@ -401,6 +401,7 @@ mod tests {
                     thinking: None,
                     tool_calls: None,
                     function_call: None,
+                    audio: None,
                 },
                 finish_reason: None,
                 logprobs: None,
@@ -424,6 +425,7 @@ mod tests {
                 thinking: None,
                 tool_calls: None,
                 function_call: None,
+                audio: None,
             },
             finish_reason: Some(FinishReason::Stop),
             logprobs: None,
